Mainly working now, with different local and remote roots.
Need to do some testing on Windows. It's all got a little bit messy
with these recent changes - didn't really anticipate that we'd want
different local and remote roots. Should be pretty obvious that the
default choice for local root is current directory, and for remote
is public_html. Oh well.