Made compatible with python2.6
There was a strange bug with empty files which didn't
change the remote modify date. There's now a new check
which, if local file is newer, will first check to see
if local file has zero size, then check to see if remote
file has zero size. If so, files are the same, even if
modification date is different.